scaladocs for RPC annotations, cosmetic refactors
