+++ commons-core/src/main/scala/com/avsystem/commons/rpc/RawRpcCompanion.scala
@@ -6,7 +6,7 @@ import com.avsystem.commons.macros.rpc.RPCMacros
 /**
   * Base trait for companion objects of raw RPC traits.
   */
-trait RawRPCCompanion[R] {
+trait RawRpcCompanion[R] extends RpcImplicitsProvider {
   type AsRawRPC[T] = AsRaw[T, R]
   type AsRealRPC[T] = AsReal[T, R]
   type AsRealRawRPC[T] = AsRealRaw[T, R]
@@ -14,6 +14,9 @@ trait RawRPCCompanion[R] {
   def materializeAsRaw[T]: AsRawRPC[T] = macro RPCMacros.rpcAsRaw[T, R]
   def materializeAsReal[T]: AsRealRPC[T] = macro RPCMacros.rpcAsReal[T, R]
   def materializeAsRealRaw[T]: AsRealRawRPC[T] = macro RPCMacros.rpcAsRealRaw[T, R]
+}
+
+trait RpcImplicitsProvider {
 
   /**
     * If you want some more implicits to be visible by RPC macros,
